backend/requirements.txt
# Language Teacher Application - LLaMA Implementation

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# AI Model Integration
transformers==4.36.2
torch==2.1.2
accelerate==0.25.0
bitsandbytes==0.41.3
sentencepiece==0.1.99
protobuf==3.20.3

# GPT Models
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
speechrecognition==3.10.0

# Natural Language Processing
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# File Handling
python-magic==0.4.27
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development
watchfiles==0.21.0
